+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 Added option to show or hide the disabled project items.
 - [#31](https://github.com/denis-shienkov/vscode-qbs/issues/31)
 Rigt now the disabled project items are marked as striked text.
+- [#2](https://github.com/denis-shienkov/vscode-qbs/issues/2)
+Implemented the project re-resolving when the qbs/js files change.
 
 ## 0.0.9